Refactored blog list layout
