+++ src/pages/index.js
@@ -1,19 +1,19 @@
 import React from 'react'
-import Layout from "../components/layout"
-import { Link, graphql } from 'gatsby'
+import { graphql } from 'gatsby'
 import { Helmet } from 'react-helmet'
 
-import Title from '../components/Title'
+import Layout from '../components/layout'
 import Bio from '../components/Bio'
+import BlogList from '../templates/blog-list-template'
+
 import typography from '../utils/typography'
 const rhythm = typography.rhythm
-const scale = typography.scale
 
 class Index extends React.Component {
   render() {
     const siteTitle = this.props.data.site.siteMetadata.title
     const posts = this.props.data.allMarkdownRemark.edges.filter(item => 
-      item.node.frontmatter.type !== "page")
+      item.node.frontmatter.type !== 'page')
 
     return (
       <Layout location={this.props.location}>
@@ -24,39 +24,7 @@ class Index extends React.Component {
             marginTop: rhythm(2),
             marginBottom: rhythm(2)
           }} />
-          {posts.map(({ node }, index) => {
-            const title = node.frontmatter.title || node.fields.slug
-            return (
-              <div key={node.fields.slug}>
-                <h2 style={{
-                  marginBottom: rhythm(0.25),
-                  fontSize: node.frontmatter.passthroughUrl ? '1.21225rem' : '1.618rem',
-                }}>
-                  <Title title={title} passthroughUrl={node.frontmatter.passthroughUrl} slug={node.fields.slug} />
-                </h2>
-                <p className="post-meta" style={{
-                  ...scale(-1 / 5),
-                  marginBottom: node.frontmatter.passthroughUrl ? rhythm(0) : rhythm(0.75),
-                  display: 'block',
-                }}>{node.frontmatter.date}</p>
-                <div style={{
-                  marginBottom: rhythm(2)
-                }}>
-                  <div dangerouslySetInnerHTML={{ __html: index > 1 ? node.excerpt : node.html }} style={{
-                    marginBottom: rhythm(0.5)
-                  }} />
-                  <p>
-                    <Link to={node.fields.slug}>
-                      ⌘
-                    </Link>
-                  </p>
-                  <hr style={{
-                    marginTop: rhythm(2)
-                  }} />
-                </div>
-              </div>
-            )
-          })}
+          <BlogList posts={posts} />
         </div>
       </Layout>
     )
